Lots of updates
This commit contains a lot of changes made. Logs data to a text file that is created and stored by the file manager within the app sandbox. Timer is implemented via a button press. Lots of UI updates to better use the app and see what's going on. Data logs only when a button is pressed, and only if theres a hand present in the frame.